pyproject.toml
[tool.poetry]
name = "drf-auth-example"
version = "0.1.0"
description = ""
authors = ["Andrew Deaver <deaver.andrew@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
